added docs and refactor
